Major refactoring update
Functionality moved among modules in the following way:

- packet extraction from ROOT and NPY files moved into utils/io_utils.py
- packet manipulation and shower line value generators moved into utils/data_utils.py
- packet template moved into utils/packets/packet_utils.py

Packet template provides functionality formerly part of packet manipulation,
namely packet attributes (frame width and height, EC module width and height,
number of frames, number of rows and columns of EC modules, packet shape)
and also unit conversions from x, y to EC x, EC y and EC idx

Packet manipulator provides functionality for drawing simulated shower lines,
simulating EC malfunctions and creating projections from packets. It depends
on packet template and provides optional checking if a packet passed to it
conforms to the template it was created with.

Packet extraction functions for NPY and ROOT files are now part of a class,
and not standalone functions. This class also depends on packet template
for checking the packet structure of the file and if it does not conform
to the template, it throws an exception. The template used can be changed
after class instantiation. Also, unlike the previous version, the extraction
functions simply extract packets, but do not assign classification labels
to them; any post-extraction processing must be done by the caller (in this
case, dataset_creator.py) via a callback function passed to the extraction
function.

All modules are covered with unit tests located in the same directory
or in a subdirectory 'test' below the directory they are in.

Also changed cmd interfaces:
- added minimum and maximum shower duration arguments to datagen
- removed -a alias for --acqfile argument from checker