Bugfix: width and height of packet data
The Shape of a packet as represented in numpy was assumed to be (num_gtu, width, height)
but actually it is (num_gtu, height, width) since an individual frame uses matrix
conventions for indexing. Width and height were swapped in every bit of code that uses
them.

Also changed code of unit tests to reflect this change.