Removed params argument from convnet_checker
Training data is rarely used for evaluation, and due to changes in data_generator,
params would not have worked anyway.

As it is, checker cannot evaluate both single input as well as multi-input convnets,
this functionality will be added later, when it is resolved in convnet_trainer as well.

Also minor bugfix in convnet_checker script.