Major refactoring (primarily data generator and utils), broken unit tests
Data generator is now a separate class that creates dataset files of simulated data
and provides functions for creating shower and noise packets (as well as properties
to fine tune things like the number of EC malfunctions per frame etc).

Simulated shower properties, such as the ranges for start coordinates, shower peak
intensity across a dataset as well as duration across a dataset have been extracted
into another class (synth_data_utils.simulated_shower_template) which also provides
random value generators for returning values (that are within thier respective range).
The class requires a packet template to check if the duration and coordinate values'
ranges are allowable with respect to (expected) packet dimensions. The class also has
as one of its properties the shower values generator to be used.

This module also contains the functions for selecting random ECs (moved here from packet
template) and for creating a simulated shower line as standalone functions. The latter
function comes as a base version and a wrapper which supplies the arguments for the base
funtion from a shower_template object passed to it).

-- cmdint/common_args --

Params_args and shower_args that used to be in utils/synth_data_utils have been removed
and their functionality moved elsewhere. Since certain arguments are expected to be used
by multiple scripts, these and the functionality for checking and transforming these
arguments into more usable forms (packet template, dataset helper etc.) have been
re-implemented as standalone functions in a new module (cmdint.common_args).

Current supported functionality includes specifying outputs during dataset creation
(raw packets or yx, gtux, gtuy projections) or equivalent inputs for consumption by
neural networks, though the latter is not yet used. This comes with args checking and
functionality for creating default filenames for each type of data (and targets). Also
supported are packet_dims which are converted into a packet template (the checking is
done by the template constructor).

-- utils/common_utils --

Contains primarily functions used to convert arguments for shower property and bad EC
intervals from a variety of supported types into a tuple of 2 integers (their internal
representation) and for checking that these interval tuples are valid.

-- utils/dataset_utils --

All numpy related functionality relating to projection creation has been moved here from
packet_manipulator (the latter was removed). This module contains a 'helper' class which
creates 'holders' (data structures in the form of numpy arrays) used to hold dataset items
of a specific type (raw or projections) relative to the source packets from which they are
created. This class essentially controls the type of items that are used for a dataset with
boolean flags implemented as properties as well as enables conversion of individual packets
into these dataset items.

Also included here: Functionality related to shuffling dataset items and persisting datasets
to secondary storage (both as standalone functions).

-- Misc --
Dataset creator was also modified: now uses output type specification command-line arguments
as well as dataset_utils to create a dataset, convert extracted packets to dataset items and
persist the new dataset.

Should be noted, that a major unit test refactoring is needed, but it would make this commit
way larger and harder to parse than it already is, so it is deferred to the nex commit. Do not
run unit tests in this commit, they are all broken now.