Major refactoring: encapsulate dataset operations into single class
Class dataset helper did not offer the best functionality, so remove it.

Create class numpy_dataset, offering functionality to load and store datasets,
add and get data items, targets and metadata and check if items of a certain
type (packets, various projection types) are in this dataset (as well as check
attributes of the set). Also store most attributes of a dataset in a config.ini
file, so that on the command line, the user has to only specify the datasets
name, directory in which its files are located and item types requested from
the dataset as neural network inputs.

Also modify command line interfaces and the semantics of arguments to reflect
this refactoring.

Also modify scripts and modules using dataset_helper (except convnet_checker)
to use this class instead.