Project reorganization
Create a common module for templates of packet (or packet dimensions),
simulated showers as well as simulated backgrounds.

Also rename data_utils to shower_generators as that is pretty much
what is in that module.

Also modify imports where necessary (except convnet_checker, as it is
still being worked out and redesigned).

Also refactor unit tests to reflect the module changes.