Finally fix convnet_checker
Make frame creation to visualize dataset data be the sole responsibility
of the dataset_visualizer script. Convnet checker reports should only use
<img> tags referencing the images of dataset items created by it. Remove
--onlyframes command-line switch, as it no longer does anything.

The output report files can be linked to the images by copying the directory
they are directly under into the directory storing the images in the following
way:

|--root_dir
|----img -- images, with subdirectories for individual item types
|           (yx, gtux, gtuy, etc.)
|----report_whatever -- report html files

No support for visualizing raw packets, as we do not yet have 3D Convnets
to test, nor a dataset for them.

Note: Should probably invent a common configuration file or module constant
to store the filename format and directories in which images are saved so
both scripts can use the same parameters - currently, both the values are
hard-coded into the scripts or their dependencies.

Also, frame creation functions and functions to transform metadata into text
to display in the images or the report tables have been moved to a separate
module.

Also, some minor edits in network_utils and dataset_utils.