Make dataset condenser allow more custom options when creating a dataset
Filght and simu are kept as standalone functions for legacy purposes.
Custom option allows setting a static target for all frames and to set
the start and end frames or GTUs for creating projections (start and
end are interpreted analogously to start and stop in python slices,
i.e. the end frame is not used). Also add more metadata to the dataset
TSv.

Also rewrite io_utils so that it returns all packets at once, rather
than calling back a handler on every packet extracted (it's easier to
implement). Unit tests for this module are currently broken fix later.