Make model classes able to use different internal and external weight formats.
The external format is shown when a caller requests the weights of trainable
layers and is accepted when assigning the values of weights of layers.

Also rewrite tests to use the external format.

Also bugfix in tests: when a test class (TestCase) for a network model class
is subclassed, make sure the test methods when called use the class of model
tested by the subclass. This is achieved by the optional model parameter in
the test method signatures.