Added layer path managment to graph builder.
A path is an ordered series of layer names that can be used to guide
ordering activations of particular layers during e.g. visualization.

A new path is started and termineted explicitly and every layer added
betweeen the start and end of the path can be added to the path unless
excluded using a method parameter (some are excluded by default, e.g.
dropout and reshape layers).